Allow clones' name/symbol to be set w/ ERC20Upgradeable
This brings in initializer functionality on the ERC20 class,
and removes the hack for storing the member variables ourselves.

The full ERC20Upgradeable feature set is not required, but it is the
most minimal way to get a verified Initializable ERC20 interface.

Note that the @openzeppelin/contracts-upgradeable package only supports
_disableInitializers from v4.6.0